rename configs to match dataset description
rather than library used to read data

i.e. xarray -> ukcp18
