+++ src/ml_downscaling_emulator/score_sde_pytorch_hja22/configs/vesde/ukcp18_cncsnpp_continuous.py
@@ -14,33 +14,28 @@
 # limitations under the License.
 
 # Lint as: python3
-"""Training NCSN++ on precip data with sub-VP SDE."""
-from configs.default_xarray_configs import get_default_configs
+"""Training NCSN++ on precip data with VE SDE."""
+from ml_downscaling_emulator.score_sde_pytorch_hja22.configs.default_ukcp18_configs import get_default_configs
 
 
 def get_config():
   config = get_default_configs()
   # training
   training = config.training
-  training.sde = 'subvpsde'
+  training.sde = 'vesde'
   training.continuous = True
-  training.reduce_mean = True
 
   # sampling
   sampling = config.sampling
   sampling.method = 'pc'
-  sampling.predictor = 'euler_maruyama'
-  sampling.corrector = 'none'
-
-  # data
-  data = config.data
-  data.centered = True
+  sampling.predictor = 'reverse_diffusion'
+  sampling.corrector = 'langevin'
 
   # model
   model = config.model
-  model.name = 'ncsnpp'
-  model.scale_by_sigma = False
-  model.ema_rate = 0.9999
+  model.name = 'cncsnpp'
+  model.scale_by_sigma = True
+  model.ema_rate = 0.999
   model.normalization = 'GroupNorm'
   model.nonlinearity = 'swish'
   model.nf = 128
@@ -57,9 +52,11 @@ def get_config():
   model.progressive_input = 'residual'
   model.progressive_combine = 'sum'
   model.attention_type = 'ddpm'
-  model.embedding_type = 'positional'
   model.init_scale = 0.
   model.fourier_scale = 16
   model.conv_size = 3
+
+  # data
+  data = config.data
 
   return config
